pyproject.toml
[project]
name = "send-water-reminder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "load-dotenv>=0.1.0",
]
